Break up models.py into separate files
