+++ ml-agents/mlagents/trainers/torch/networks.py
@@ -1,17 +1,16 @@
-from enum import Enum
-from typing import Callable, NamedTuple, List, Optional, Dict, Tuple
+from typing import Callable, NamedTuple, List, Dict, Tuple
 
 import torch
 from torch import nn
-import numpy as np
 
 from mlagents_envs.base_env import ActionType
 from mlagents.trainers.torch.distributions import (
     GaussianDistribution,
     MultiCategoricalDistribution,
 )
-from mlagents.trainers.exception import UnityTrainerException
-from mlagents.trainers.settings import EncoderType, NetworkSettings
+from mlagents.trainers.settings import NetworkSettings
+from mlagents.trainers.torch.utils import ModelUtils
+from mlagents.trainers.torch.decoders import ValueHeads
 
 ActivationFunction = Callable[[torch.Tensor], torch.Tensor]
 EncoderFunction = Callable[
@@ -21,53 +20,10 @@ EncoderFunction = Callable[
 EPSILON = 1e-7
 
 
-def list_to_tensor(
-    ndarray_list: List[np.ndarray], dtype: Optional[torch.dtype] = None
-) -> torch.Tensor:
-    """
-    Converts a list of numpy arrays into a tensor. MUCH faster than
-    calling as_tensor on the list directly.
-    """
-    return torch.as_tensor(np.asanyarray(ndarray_list), dtype=dtype)
-
-
-class LearningRateSchedule(Enum):
-    CONSTANT = "constant"
-    LINEAR = "linear"
-
-
 class NormalizerTensors(NamedTuple):
     steps: torch.Tensor
     running_mean: torch.Tensor
     running_variance: torch.Tensor
-
-
-def break_into_branches(
-    concatenated_logits: torch.Tensor, action_size: List[int]
-) -> List[torch.Tensor]:
-    """
-    Takes a concatenated set of logits that represent multiple discrete action branches
-    and breaks it up into one Tensor per branch.
-    :param concatenated_logits: Tensor that represents the concatenated action branches
-    :param action_size: List of ints containing the number of possible actions for each branch.
-    :return: A List of Tensors containing one tensor per branch.
-    """
-    action_idx = [0] + list(np.cumsum(action_size))
-    branched_logits = [
-        concatenated_logits[:, action_idx[i] : action_idx[i + 1]]
-        for i in range(len(action_size))
-    ]
-    return branched_logits
-
-
-def actions_to_onehot(
-    discrete_actions: torch.Tensor, action_size: List[int]
-) -> List[torch.Tensor]:
-    onehot_branches = [
-        torch.nn.functional.one_hot(_act.T, action_size[i])
-        for i, _act in enumerate(discrete_actions.T)
-    ]
-    return onehot_branches
 
 
 class NetworkBody(nn.Module):
@@ -413,139 +369,6 @@ class Critic(nn.Module):
         return self.value_heads(embedding)
 
 
-class Normalizer(nn.Module):
-    def __init__(self, vec_obs_size, **kwargs):
-        super().__init__(**kwargs)
-        self.normalization_steps = torch.tensor(1)
-        self.running_mean = torch.zeros(vec_obs_size)
-        self.running_variance = torch.ones(vec_obs_size)
-
-    def forward(self, inputs):
-        normalized_state = torch.clamp(
-            (inputs - self.running_mean)
-            / torch.sqrt(self.running_variance / self.normalization_steps),
-            -5,
-            5,
-        )
-        return normalized_state
-
-    def update(self, vector_input):
-        steps_increment = vector_input.size()[0]
-        total_new_steps = self.normalization_steps + steps_increment
-
-        input_to_old_mean = vector_input - self.running_mean
-        new_mean = self.running_mean + (input_to_old_mean / total_new_steps).sum(0)
-
-        input_to_new_mean = vector_input - new_mean
-        new_variance = self.running_variance + (
-            input_to_new_mean * input_to_old_mean
-        ).sum(0)
-        self.running_mean = new_mean
-        self.running_variance = new_variance
-        self.normalization_steps = total_new_steps
-
-    def copy_from(self, other_normalizer: "Normalizer") -> None:
-        self.normalization_steps.data.copy_(other_normalizer.normalization_steps.data)
-        self.running_mean.data.copy_(other_normalizer.running_mean.data)
-        self.running_variance.copy_(other_normalizer.running_variance.data)
-
-
-class ValueHeads(nn.Module):
-    def __init__(self, stream_names, input_size, output_size=1):
-        super().__init__()
-        self.stream_names = stream_names
-        _value_heads = {}
-
-        for name in stream_names:
-            value = nn.Linear(input_size, output_size)
-            _value_heads[name] = value
-        self.value_heads = nn.ModuleDict(_value_heads)
-
-    def forward(self, hidden):
-        value_outputs = {}
-        for stream_name, head in self.value_heads.items():
-            value_outputs[stream_name] = head(hidden).squeeze(-1)
-        return (
-            value_outputs,
-            torch.mean(torch.stack(list(value_outputs.values())), dim=0),
-        )
-
-
-class VectorEncoder(nn.Module):
-    def __init__(self, input_size, hidden_size, num_layers, **kwargs):
-        super().__init__(**kwargs)
-        self.layers = [nn.Linear(input_size, hidden_size)]
-        for _ in range(num_layers - 1):
-            self.layers.append(nn.Linear(hidden_size, hidden_size))
-            self.layers.append(nn.ReLU())
-        self.seq_layers = nn.Sequential(*self.layers)
-
-    def forward(self, inputs):
-        return self.seq_layers(inputs)
-
-
-def conv_output_shape(h_w, kernel_size=1, stride=1, pad=0, dilation=1):
-    from math import floor
-
-    if type(kernel_size) is not tuple:
-        kernel_size = (kernel_size, kernel_size)
-    h = floor(
-        ((h_w[0] + (2 * pad) - (dilation * (kernel_size[0] - 1)) - 1) / stride) + 1
-    )
-    w = floor(
-        ((h_w[1] + (2 * pad) - (dilation * (kernel_size[1] - 1)) - 1) / stride) + 1
-    )
-    return h, w
-
-
-def pool_out_shape(h_w, kernel_size):
-    height = (h_w[0] - kernel_size) // 2 + 1
-    width = (h_w[1] - kernel_size) // 2 + 1
-    return height, width
-
-
-class SimpleVisualEncoder(nn.Module):
-    def __init__(self, height, width, initial_channels, output_size):
-        super().__init__()
-        self.h_size = output_size
-        conv_1_hw = conv_output_shape((height, width), 8, 4)
-        conv_2_hw = conv_output_shape(conv_1_hw, 4, 2)
-        self.final_flat = conv_2_hw[0] * conv_2_hw[1] * 32
-
-        self.conv1 = nn.Conv2d(initial_channels, 16, [8, 8], [4, 4])
-        self.conv2 = nn.Conv2d(16, 32, [4, 4], [2, 2])
-        self.dense = nn.Linear(self.final_flat, self.h_size)
-
-    def forward(self, visual_obs):
-        conv_1 = torch.relu(self.conv1(visual_obs))
-        conv_2 = torch.relu(self.conv2(conv_1))
-        # hidden = torch.relu(self.dense(conv_2.view([-1, self.final_flat])))
-        hidden = torch.relu(self.dense(torch.reshape(conv_2, (-1, self.final_flat))))
-        return hidden
-
-
-class NatureVisualEncoder(nn.Module):
-    def __init__(self, height, width, initial_channels, output_size):
-        super().__init__()
-        self.h_size = output_size
-        conv_1_hw = conv_output_shape((height, width), 8, 4)
-        conv_2_hw = conv_output_shape(conv_1_hw, 4, 2)
-        conv_3_hw = conv_output_shape(conv_2_hw, 3, 1)
-        self.final_flat = conv_3_hw[0] * conv_3_hw[1] * 64
-
-        self.conv1 = nn.Conv2d(initial_channels, 32, [8, 8], [4, 4])
-        self.conv2 = nn.Conv2d(32, 64, [4, 4], [2, 2])
-        self.conv3 = nn.Conv2d(64, 64, [3, 3], [1, 1])
-        self.dense = nn.Linear(self.final_flat, self.h_size)
-
-    def forward(self, visual_obs):
-        conv_1 = torch.relu(self.conv1(visual_obs))
-        conv_2 = torch.relu(self.conv2(conv_1))
-        conv_3 = torch.relu(self.conv3(conv_2))
-        hidden = torch.relu(self.dense(conv_3.view([-1, self.final_flat])))
-        return hidden
-
-
 class GlobalSteps(nn.Module):
     def __init__(self):
         super().__init__()
@@ -560,135 +383,3 @@ class LearningRate(nn.Module):
         # Todo: add learning rate decay
         super().__init__()
         self.learning_rate = torch.Tensor([lr])
-
-
-class ResNetVisualEncoder(nn.Module):
-    def __init__(self, height, width, initial_channels, final_hidden):
-        super().__init__()
-        n_channels = [16, 32, 32]  # channel for each stack
-        n_blocks = 2  # number of residual blocks
-        self.layers = []
-        last_channel = initial_channels
-        for _, channel in enumerate(n_channels):
-            self.layers.append(
-                nn.Conv2d(last_channel, channel, [3, 3], [1, 1], padding=1)
-            )
-            self.layers.append(nn.MaxPool2d([3, 3], [2, 2]))
-            height, width = pool_out_shape((height, width), 3)
-            for _ in range(n_blocks):
-                self.layers.append(self.make_block(channel))
-            last_channel = channel
-        self.layers.append(nn.ReLU())
-        self.dense = nn.Linear(n_channels[-1] * height * width, final_hidden)
-
-    @staticmethod
-    def make_block(channel):
-        block_layers = [
-            nn.ReLU(),
-            nn.Conv2d(channel, channel, [3, 3], [1, 1], padding=1),
-            nn.ReLU(),
-            nn.Conv2d(channel, channel, [3, 3], [1, 1], padding=1),
-        ]
-        return block_layers
-
-    @staticmethod
-    def forward_block(input_hidden, block_layers):
-        hidden = input_hidden
-        for layer in block_layers:
-            hidden = layer(hidden)
-        return hidden + input_hidden
-
-    def forward(self, visual_obs):
-        batch_size = visual_obs.shape[0]
-        hidden = visual_obs
-        for layer in self.layers:
-            if isinstance(layer, nn.Module):
-                hidden = layer(hidden)
-            elif isinstance(layer, list):
-                hidden = self.forward_block(hidden, layer)
-        before_out = hidden.view(batch_size, -1)
-        return torch.relu(self.dense(before_out))
-
-
-class ModelUtils:
-    # Minimum supported side for each encoder type. If refactoring an encoder, please
-    # adjust these also.
-    MIN_RESOLUTION_FOR_ENCODER = {
-        EncoderType.SIMPLE: 20,
-        EncoderType.NATURE_CNN: 36,
-        EncoderType.RESNET: 15,
-    }
-
-    @staticmethod
-    def swish(input_activation: torch.Tensor) -> torch.Tensor:
-        """Swish activation function. For more info: https://arxiv.org/abs/1710.05941"""
-        return torch.mul(input_activation, torch.sigmoid(input_activation))
-
-    @staticmethod
-    def get_encoder_for_type(encoder_type: EncoderType) -> nn.Module:
-        ENCODER_FUNCTION_BY_TYPE = {
-            EncoderType.SIMPLE: SimpleVisualEncoder,
-            EncoderType.NATURE_CNN: NatureVisualEncoder,
-            EncoderType.RESNET: ResNetVisualEncoder,
-        }
-        return ENCODER_FUNCTION_BY_TYPE.get(encoder_type)
-
-    @staticmethod
-    def _check_resolution_for_encoder(
-        vis_in: torch.Tensor, vis_encoder_type: EncoderType
-    ) -> None:
-        min_res = ModelUtils.MIN_RESOLUTION_FOR_ENCODER[vis_encoder_type]
-        height = vis_in.shape[1]
-        width = vis_in.shape[2]
-        if height < min_res or width < min_res:
-            raise UnityTrainerException(
-                f"Visual observation resolution ({width}x{height}) is too small for"
-                f"the provided EncoderType ({vis_encoder_type.value}). The min dimension is {min_res}"
-            )
-
-    @staticmethod
-    def create_encoders(
-        observation_shapes: List[Tuple[int, ...]],
-        h_size: int,
-        num_layers: int,
-        vis_encode_type: EncoderType,
-        action_size: int = 0,
-    ) -> Tuple[nn.ModuleList, nn.ModuleList, nn.ModuleList]:
-        """
-        Creates visual and vector encoders, along with their normalizers.
-        :param observation_shapes: List of Tuples that represent the action dimensions.
-        :param action_size: Number of additional un-normalized inputs to each vector encoder. Used for
-            conditioining network on other values (e.g. actions for a Q function)
-        :param h_size: Number of hidden units per layer.
-        :param num_layers: Depth of MLP per encoder.
-        :param vis_encode_type: Type of visual encoder to use.
-        :return: Tuple of visual encoders, vector encoders, and vector normalizers, each as a list.
-        """
-        visual_encoders: List[nn.Module] = []
-        vector_encoders: List[nn.Module] = []
-        vector_normalizers: List[nn.Module] = []
-
-        visual_encoder_class = ModelUtils.get_encoder_for_type(vis_encode_type)
-        vector_size = 0
-        for i, dimension in enumerate(observation_shapes):
-            if len(dimension) == 3:
-                visual_encoders.append(
-                    visual_encoder_class(
-                        dimension[0], dimension[1], dimension[2], h_size
-                    )
-                )
-            elif len(dimension) == 1:
-                vector_size += dimension[0]
-            else:
-                raise UnityTrainerException(
-                    f"Unsupported shape of {dimension} for observation {i}"
-                )
-        vector_normalizers.append(Normalizer(vector_size))
-        vector_encoders.append(
-            VectorEncoder(vector_size + action_size, h_size, num_layers)
-        )
-        return (
-            nn.ModuleList(visual_encoders),
-            nn.ModuleList(vector_encoders),
-            nn.ModuleList(vector_normalizers),
-        )
